add files of my homework (read description)
In this homework I should to download hierarchy.json from server, according to the included info dynamically create an nested list, buttons "Download" , "Clear", table and id(n).json files as we have in hierarchy.json.
When u click on button "Download", a request is sent to the server and hierarchy.json file is returned. Nested list is created dynamically.
Futher when u click on item of list, a request is sent to the server and the corresponding id(n).json file is returned. Element of list become marked and data from file send to the table. Also this data is saved in local storage with own id (as same as file's name).
If u click on button "Clear" all data from table and all records from local srorage will be deleted and pushed items of list will get their common state.